Extract single vote component
